1. fix the tensor shape bug reported in issue https://github.com/nunchaku-tech/deepcompressor/issues/89 2. add sdxl support in nunchaku backend 3. add sdxl-turbo yaml config in examples
